Move PO types into seperate folder
