+++ src/index.ts
@@ -6,10 +6,7 @@ export * from "./currency";
 export * from "./invoice";
 export * from "./membership";
 export * from "./product";
-export * from "./purchase-order-item";
-export * from "./purchase-order-receive";
 export * from "./purchase-order";
-export * from "./purchase-tax";
 export * from "./sales-order";
 export * from "./shipping-policy";
 export * from "./store";
