+++ src/purchase-order/comment.ts
@@ -1,0 +1,11 @@
+export interface PurchaseOrderComment {
+  commentId: string;
+  commentType: string;
+  date: Date;
+  dateDescription: string;
+  description: string;
+  operationType: string;
+  purchaseorderId: string;
+  time: string;
+  transactionType: string;
+}
