+++ tsconfig.json
@@ -2,7 +2,8 @@
   "compilerOptions": {
     "target": "ES2020",
     "outDir": "./build",
-    "baseUrl": "./src"
+    "baseUrl": "./src",
+    "moduleResolution": "Node"
   },
   "exclude": ["node_modules", "test"]
 }
